More style fixes
