+++ app/src/main/java/com/wndenis/snipsnap/calendar/ScheduleCalendar.kt
@@ -40,13 +40,13 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.vanpra.composematerialdialogs.color.ColorPalette
-import com.wndenis.snipsnap.calendar.calcomponents.DayDividers
-import com.wndenis.snipsnap.calendar.calcomponents.DaysRow
-import com.wndenis.snipsnap.calendar.calcomponents.HoursRow
-import com.wndenis.snipsnap.calendar.calcomponents.MonthDividers
-import com.wndenis.snipsnap.calendar.calcomponents.MonthRow
-import com.wndenis.snipsnap.calendar.calcomponents.YearDividers
-import com.wndenis.snipsnap.calendar.calcomponents.YearRow
+import com.wndenis.snipsnap.calendar.components.DayDividers
+import com.wndenis.snipsnap.calendar.components.DaysRow
+import com.wndenis.snipsnap.calendar.components.HoursRow
+import com.wndenis.snipsnap.calendar.components.MonthDividers
+import com.wndenis.snipsnap.calendar.components.MonthRow
+import com.wndenis.snipsnap.calendar.components.YearDividers
+import com.wndenis.snipsnap.calendar.components.YearRow
 import com.wndenis.snipsnap.data.CalendarAdapter
 import com.wndenis.snipsnap.data.CalendarEvent
 import com.wndenis.snipsnap.data.CalendarSection
