+++ app/src/main/java/com/wndenis/snipsnap/calendar/components/LocalDateTimeData.kt
@@ -1,0 +1,11 @@
+package com.wndenis.snipsnap.calendar.components
+
+import androidx.compose.ui.layout.ParentDataModifier
+import androidx.compose.ui.unit.Density
+import java.time.LocalDateTime
+
+internal data class LocalDateTimeData(
+    val localDateTime: LocalDateTime,
+) : ParentDataModifier {
+    override fun Density.modifyParentData(parentData: Any?) = this@LocalDateTimeData
+}
